refactor comments and class names
