+++ app/models/documentsModel.js
@@ -42,4 +42,4 @@ export default function (sequelize, Sequelize) {
       }
     });
   return Document;
-};
+}
