+++ app/models/rolesModel.js
@@ -15,4 +15,4 @@ export default function (sequelize, DataTypes) {
     }
   });
   return Role;
-};
+}
