+++ app/models/usersModel.js
@@ -57,4 +57,4 @@ export default function (sequelize, Sequelize) {
   });
 
   return User;
-};
+}
