+++ app/routes/index.js
@@ -3,8 +3,8 @@ import express from 'express';
 const homeRoute = express.Router();
 
 /**
-  * Route for the homepage.
-  */
+ * Route for the homepage.
+ */
 homeRoute.get('/', (req, res) => {
   res.json({ message: 'Welcome to Document Management App.' });
 });
