+++ app/routes/roles.js
@@ -1,27 +1,27 @@
 import express from 'express';
-import RoleController from '../controllers/roleController';
+import RoleController from '../controllers/RoleController';
 
 const roleControl = new RoleController();
 const roleRoutes = express.Router();
 
 /**
-  * Route for creating roles.
-  */
+ * Route for creating roles.
+ */
 roleRoutes.post('/', roleControl.createRole);
 
 /**
-  * Route for getting roles.
-  */
+ * Route for getting roles.
+ */
 roleRoutes.get('/', roleControl.getRoles);
 
 /**
-  * Route for updating roles.
-  */
+ * Route for updating roles.
+ */
 roleRoutes.put('/:id', roleControl.updateRole);
 
 /**
-  * Route for deleting roles.
-  */
+ * Route for deleting roles.
+ */
 roleRoutes.delete('/:id', roleControl.deleteRole);
 
 export default roleRoutes;
